pyproject.toml
[project]
name = "andy-protogen"
version = "0.1.0"
description = "Protogen electronic visor control system for Raspberry Pi 5 + HUB75 LED matrices"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rpi = [
    "adafruit-blinka-raspberry-pi5-piomatter>=1.0.0",
    "gpiod>=2.2",
]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
]
mock = [
    "pygame-ce>=2.5",
]
monitoring = [
    "psutil>=5.9.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
all = [
    "andy-protogen[rpi,web,monitoring,dev]",
]

[project.scripts]
protogen = "protogen.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
